requirements/base.txt
Django==1.9.2
whitenoise==2.0.6
django-environ==0.4.0
django-model-utils